Many Snake Changes, Minor Rectangle Change
For Rectangle: added setter methods for x and y
For Snake: Made a constructor to initialize with a size, direction, and
x and y coordinates. Also added: a default constructor with no
parameters, a system for checking when the snake is alive or not, checks
for non-valid turns (trying to go backwards).